requirements.txt
# ===============================
# FastAPI Framework
# ===============================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ===============================
# Machine Learning
# ===============================
tensorflow==2.15.0
numpy==1.26.4

# ===============================
# Image Processing
# ===============================
pillow==10.2.0

# ===============================
# Testing
# ===============================
pytest==8.1.1
pytest-mock==3.12.0
httpx==0.27.0

# ===============================
# Environment Management
# ===============================
python-dotenv==1.0.1
